+++ examples/qt_sdl/user/Virtual_TFT_Port.h
@@ -48,7 +48,7 @@ void vtFillMultipleColors(int32_t x1, int32_t y1, int32_t x2, int32_t y2, color_
 void vtSetPoint(int32_t x, int32_t y, color_typedef color);
 color_typedef vtGetPoint(int32_t x, int32_t y);
 bool vtMouseGetPoint(int16_t *x,int16_t *y);
-bool vtGetKeyState(size_t value);
+bool vtGetKeyState(uint16_t value);
 
 #ifdef __cplusplus
 } /* extern "C" */
